Add Stat schema plugin

+++ spec/shared/filesystem_setup.rb
@@ -4,7 +4,7 @@ RSpec.shared_context 'filesystem setup' do
   let(:uri) { SPEC_ROOT.join('../tmp/test') }
 
   before do
-    uri.rmtree
+    uri.rmtree if uri.directory?
     uri.mkpath
   end
 end
